+++ client/src/pages/myPage/myPageEdit/MyPageEdit.js
@@ -1,8 +1,9 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import MyPageHeader from "./MyPageHeader";
-import ModalConfirm from "../../components/ui/modals/ModalConfirm";
-import Validations from "../register/Validations";
+import MyPageHeader from "../MyPageHeader";
+import MyPageNav from "../MyPageNav";
+import ModalConfirm from "../../../components/ui/modals/ModalConfirm";
+import Validations from "../../register/Validations";
 import axios from "axios";
 import { useNavigate, useParams, Link } from "react-router-dom";
 import { useForm } from "react-hook-form";
@@ -74,26 +75,9 @@ const MyPageEdit = () => {
 
   return (
     <>
-      <MFContainer>
+      <MEContainer>
         <MyPageHeader />
-        <BtnContainer>
-          <Link to="/mypage/favorite">
-            <button>관심목록</button>
-          </Link>
-          <Link to="/mypage/mypost">
-            <button>내가 쓴 게시글</button>
-          </Link>
-          <Link to="/chat/list/:id">
-            <button>채팅 목록</button>
-          </Link>
-          <Link to="/mypage/edit">
-            <button className="active">회원정보 수정</button>
-          </Link>
-          <Link to="/mypage/signout">
-            <button>회원 탈퇴</button>
-          </Link>
-        </BtnContainer>
-
+        <MyPageNav />
         <EditContainer>
           <EditContent>
             <form onSubmit={handleSubmit(onSubmit)}>
@@ -154,55 +138,19 @@ const MyPageEdit = () => {
             </form>
           </EditContent>
         </EditContainer>
-      </MFContainer>
+      </MEContainer>
     </>
   );
 };
 
 export default MyPageEdit;
 
-const MFContainer = styled.div`
+const MEContainer = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
   width: 100vw;
   height: 100%;
-`;
-
-const BtnContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 53.75rem;
-  height: 3.75rem;
-  margin-bottom: 1.8125rem;
-
-  button {
-    width: 8.75rem;
-    height: 3.75rem;
-    border-radius: 0.625rem;
-    border: 0.125rem solid ${(props) => props.theme.gray4};
-    background-color: ${(props) => props.theme.bgColor};
-    font-size: 1.125rem;
-    font-family: "NotoSansKR-Bold";
-    color: ${(props) => props.theme.gray3};
-    :hover {
-      background-color: ${(props) => props.theme.primary};
-      border: none;
-      color: ${(props) => props.theme.white};
-    }
-    &:focus {
-      background-color: ${(props) => props.theme.primary};
-      border: none;
-      color: ${(props) => props.theme.white};
-    }
-  }
-
-  .active {
-    background-color: ${(props) => props.theme.primary};
-    border: none;
-    color: ${(props) => props.theme.white};
-  }
 `;
 
 const EditContainer = styled.div`
